+++ OpenDreamShared/Network/Messages/MsgSound.cs
@@ -14,7 +14,7 @@ namespace OpenDreamShared.Network.Messages {
 
         public ushort Channel;
         public ushort Volume;
-        public int Offset;
+        public float Offset;
         public int? ResourceId;
         public FormatType? Format; // TODO: This should probably be sent along with the sound resource instead somehow
         //TODO: Frequency and friends
@@ -22,7 +22,7 @@ namespace OpenDreamShared.Network.Messages {
         public override void ReadFromBuffer(NetIncomingMessage buffer, IRobustSerializer serializer) {
             Channel = buffer.ReadUInt16();
             Volume = buffer.ReadUInt16();
-            Offset = buffer.ReadInt32();
+            Offset = buffer.ReadFloat();
 
             if (buffer.ReadBoolean()) {
                 ResourceId = buffer.ReadInt32();
